requirements.txt
Chameleon==2.16
Mako==1.0.0
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.4
WebOb==1.4
pyramid==1.5.1
pyramid-chameleon==0.1
pyramid-debugtoolbar==2.1
pyramid-layout==0.9
pyramid-mako==1.0.2
pyramid-tm==0.7
repoze.lru==0.6
transaction==1.4.3
translationstring==1.1
venusian==1.0a8
waitress==0.8.9
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.5
